Create Fractal_Tree.py
Fractal Tree Generator with use of turtle graphics library. It uses recursion to draw a tree that grows branches and sub-branches, representing fractals in nature. The turtle library makes it easy to visualize this recursive drawing, as it simulates a turtle that moves around the screen, leaving a trail as it goes.